+++ cli/src/ui/commands/AgentsMCP.tsx
@@ -8,6 +8,7 @@ import React, { useEffect, useState } from "react";
 import { AgentsMcpService } from "../../mcp/servers/mcpServer.js";
 import { useClearTerminalOnMount } from "../../utils/hooks/use_clear_terminal_on_mount.js";
 import AgentSelector from "../components/AgentSelector.js";
+import { HttpMcpTransport } from "../../mcp/types/transports/http.js";
 
 type AgentConfiguration =
   GetAgentConfigurationsResponseType["agentConfigurations"][number];
@@ -27,22 +28,28 @@ const AgentsMCP: FC<AgentsMCPProps> = ({ port, sId: requestedSIds }) => {
   const [copied, setCopied] = useState(false);
   const [mcpService, setMcpService] = useState<AgentsMcpService | null>(null);
 
-  useClearTerminalOnMount();
+  // useClearTerminalOnMount();
 
   // This useEffect handles starting the server after confirmedSelection is set
   useEffect(() => {
     if (confirmedSelection && !isServerStarted) {
-      const service = new AgentsMcpService(confirmedSelection, port);
+      const service = new AgentsMcpService(
+        new HttpMcpTransport(port),
+        confirmedSelection
+      );
       setMcpService(service);
 
-      void service
-        .startServer((url) => {
+      const start = async () => {
+        try {
+          const serverUrl = await service.startServer();
           setIsServerStarted(true);
-          setServerUrl(url);
-        })
-        .catch((err) => {
+          setServerUrl(serverUrl);
+        } catch (err: any) {
           setError(err.message || "Failed to start MCP server");
-        });
+        }
+      };
+
+      start();
     }
   }, [confirmedSelection, isServerStarted, port]);
